refactor!: Introduce Slugify trait
Resolves #4 and #5.

This is a breaking change since it removes the `Heading::anchor`.
But that's a good thing anyway since the method previously returned
wrong anchors on duplicate headings (see #4).
